LayoutTests/fast/events/drop-handler-should-not-stop-navigate.html

<!DOCTYPE html>
<html>
<body>
<div>This tests that a drop handler's default action must be prevented in order to stop navigation.
Otherwise, if event.preventDefault() is not called, navigation should occur. To test manually,
simply drag and drop another link or HTML file on this page. If navigation occurs, then the test
passed.</div>
<script>
function log(text)
{
    document.body.appendChild(document.createElement('br'));
    document.body.appendChild(document.createElement('div').appendChild(document.createTextNode(text)));
}
if (!window.layoutTestController)
    return;
window.addEventListener('beforeunload', function ()
{
    log('PASS');
    layoutTestController.notifyDone();
});
document.body.addEventListener('dragenter', function (event)
{
    log('Cancelling dragenter');
    event.preventDefault();
});
document.body.addEventListener('dragover', function (event)
{
    log('Cancelling dragover');
    event.preventDefault();
});
document.body.addEventListener('drop', function (event)
{
    log('Not preventing default event on drop.');
});
layoutTestController.dumpAsText();
layoutTestController.waitUntilDone();
log('Starting test');
eventSender.beginDragWithFiles(['DRTFakeFile']);
eventSender.mouseMoveTo(document.body.offsetLeft + 10, document.body.offsetTop + 10);
eventSender.mouseUp();
log('FAIL');
layoutTestController.notifyDone();
</script>
</body>
</html>
